Focus next or previous field when removing a field
Set the input focus to the next visible field if there is one,
or the closest previous one if there's one of those,
and otherwise, just leave it with the default focus of
the button that's been clicked.

Also remove out-of-date 'enabled' attribute of ParamField,
as its purpose is completely served by the existing
'visible' attribute of FieldLayout (its parent).

Bug: T198926
Change-Id: I4a14076889cbcc7be07e03452a46e7847823a112