Listen to click event instead of overriding 'onClick'
The 'onClick' method responds specifically to click event. In
the OO.ui.mixin.ButtonElement, it then emits 'click' event.

However, there are other ways where a click-like event happens
in the element, like pressing 'space' or 'enter' when the focus
is on the button. If we want to account for those, we should
instead listen to the internal 'click' event and use that to
activate the state toggles rather than specifically the onClick
method.

Bug: T203287
Change-Id: I7552aff8aeb5bb3beadaf54c403ebe696088aaa7