Fixing template description text, part 2
The description text is currently a light grey color. When a template is
highlighted for selection (or hovered over), it's background turns blue-
grey. The grey text on blue-grey is a bit difficult to read due to the
low contrast. Making description text slightly darker (next darkest LESS
variable), which makes it more readable but still distinguishes it from
the Template title.

Change-Id: Ib27d542f507a98f5bc68c788c5a0b759fedcc9a8