New event semantics for performer fields
This is consistent with newer schemas: we wrap in a "performer"
object, and don't send user fields for anonymous users.  This is more
meaningful than sending "null" or "0" as legacy schemas do.

Also sends the bucketed edit count.  Note that our buckets are
slightly different than the usual [0], [1-4], [5-99].  Instead, the
low range is [0-10], [11-99].  Other intervals are the same.

Bug: T262209
Bug: T269986
Depends-On: I8b35da81ab8548ec39d423bd843edbf8c97b51e8
Change-Id: Id60b87a1811ede8e3e4757f282d7792dbb9efae2