Move additional prefixsearch result to the top
This is mostly a cleanup that does not change how the code
behaves. With one exception: Setting the index to 0 moves the
prefixsearch result to the top. This matches better what users
expect. Especially expert users expect the suggester to behave
similar to the suggester they know from the top of the
MediaWiki interface.

The template search result is a combination of 9 or 10
CirrusSearch results, plus a single prefixsearch result. This
only makes a difference if the top prefixsearch result wasn't
already found by CirrusSearch. This should be rare, but is
critical for expert users.

Change-Id: I18840caa1d9c1d8f64af3044a66174820a84cf98