Fill template search results with prefix matches when Cirrus fails
This code is optimized for two relevant use cases:
1. When Cirrus found 10 results, we still want to search for the top-1
   prefix match. It might be something like the template !!. This
   should appear at the top.
2. When Cirrus found nothing, search for 10 prefix matches instead.

The code can also handle anything in between. We always search for 10
prefix matches to have the guarantee we have 10 unique results. We add
as many non-duplicates as needed, stopping when we have 10 in total.
As a result:
* The top-1 prefix match is guaranteed to be in the result set.
* All CirrusSearch matches are guaranteed to be in the result (rarely
  the 10th might get truncated).

Also see Ib047179.

Bug: T303524
Change-Id: If097d364786cf151d0e29bb49dcbfee1c2769829