Handle escape key for exiting the template form
Close the template form when pressing escape, returning to the
search form. When on the search form, escape closes the dialog.

This ensures that the warning is shown when there's data in the
form, regardless of the method used to exit the form (i.e. the
back button, or escape).

Also keep the same flow working after subsequent template forms
are opened (rather than only the first closing of the form
showing a warning about existing data).

Bug: T392442
Change-Id: Ie1047092840e016b20f19303996446e6d6030ebd